src/requirements.txt
geopandas
psycopg2-binary
python-dotenv
pandas